center position tidyup
Can't cache the GUI details as the process of adding it to the game manipulates all of the GUI Element entries in it. And setting every ones style would be a code maintenance nightmare.
Stop the close button stretching as with multiple center types it went odd.
Fix many centered entries layout and stretching.
Made the readme examples cover all the feature better.